requirements.txt
Django==2.0.1
psycopg2==2.7.1
gunicorn==19.7.1
whitenoise==3.3.1
django-extensions==1.9.9
ipdb==0.10.3
dj-database-url==0.4.2
django-reversion==2.0.13
django-model-utils==3.1.1
django-phonenumber-field==2.0.0
djangorestframework==3.7.7
django-cors-headers==2.1.0
requests==2.18.4
pq==1.5
django-fsm==2.6.0
